# openapi.yaml
openapi: 3.0.3
info:
  title: Aurora Backend API
  description: |
    Backend API untuk **Aurora Prediction Market** app.
    
    Dibangun dengan Express.js + TypeScript + PostgreSQL + Prisma ORM.
    Data market diambil dari **Polymarket API** (Gamma API, CLOB API, Data API).
    
    ## Autentikasi
    Sebagian besar endpoint user membutuhkan **JWT Bearer Token**.
    Login via `POST /api/auth/google` untuk mendapatkan token.
    
    ## Demo Mode
    Deposit dan Withdrawal bersifat **simulated** — tidak ada transaksi uang nyata.
    Setiap user baru mendapat $1000 USDC demo balance.
  version: 1.0.0
  contact:
    name: Aurora App
servers:
  - url: http://localhost:3000
    description: Local Development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT token dari POST /api/auth/google"
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Something went wrong"
    ValidationError:
      type: object
      properties:
        error:
          type: string
          example: "Validation error"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string

    # ── User ─────────────────────────────────────────────────────────────────
    User:
      type: object
      properties:
        id:
          type: string
          example: "clx1234abcd"
        email:
          type: string
          format: email
          example: "user@gmail.com"
        name:
          type: string
          example: "Aurora User"
        photoUrl:
          type: string
          nullable: true
          example: "https://lh3.googleusercontent.com/photo.jpg"
        balance:
          type: string
          example: "1000.000000"
        walletAddress:
          type: string
          nullable: true
          example: "0x742d35Cc6634C0532925a3b8D4C9B5D1b2c3D4E5"
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT token — simpan di AsyncStorage Aurora app
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        user:
          $ref: '#/components/schemas/User'

    # ── Market ───────────────────────────────────────────────────────────────
    Market:
      type: object
      properties:
        id:
          type: string
          description: conditionId dari Polymarket
          example: "0x1234abcd..."
        polymarketId:
          type: string
          example: "market-slug-123"
        slug:
          type: string
          example: "will-btc-reach-100k-2026"
        question:
          type: string
          example: "Will Bitcoin reach $100K by end of 2026?"
        description:
          type: string
          nullable: true
        category:
          type: string
          enum: [Trending, New, Politics, Sports, Crypto]
          example: "Crypto"
        tags:
          type: array
          items:
            type: string
          example: ["bitcoin", "crypto", "price"]
        outcomes:
          type: array
          items:
            type: string
          example: ["Yes", "No"]
        outcomePrices:
          type: object
          description: Probabilitas per outcome (0.0 – 1.0)
          example:
            Yes: 0.65
            No: 0.35
        tokens:
          type: array
          items:
            type: object
            properties:
              token_id:
                type: string
              outcome:
                type: string
        volume:
          type: string
          example: "1250000.50"
        volume24h:
          type: string
          example: "45000.00"
        liquidity:
          type: string
          example: "320000.00"
        active:
          type: boolean
          example: true
        closed:
          type: boolean
          example: false
        imageUrl:
          type: string
          nullable: true
        endDate:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    PricePoint:
      type: object
      properties:
        value:
          type: number
          description: Harga dalam persen (0–100)
          example: 65.32
        timestamp:
          type: integer
          description: Unix timestamp
          example: 1735689600
        label:
          type: string
          description: Label untuk chart
          example: "Jan 1"

    # ── Orderbook ────────────────────────────────────────────────────────────
...
paths:
...
  /api/auth/google:
    post:
      tags: [Auth]
      summary: Login via Google
      description: |
        Aurora app kirim data dari Google Sign In SDK.
        Backend simpan/update user di DB dan return JWT token.
        User baru otomatis mendapat $1000 USDC demo balance.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [googleId, email, name]
              properties:
                googleId:
                  type: string
                  example: "google-uid-123456789"
                email:
                  type: string
                  format: email
                  example: "user@gmail.com"
                name:
                  type: string
                  example: "Aurora User"
                photoUrl:
                  type: string
                  nullable: true
                  example: "https://lh3.googleusercontent.com/photo.jpg"
      responses:
        '200':
          description: Login berhasil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Get Current User
      description: Verifikasi JWT dan return data user yang sedang login
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Data user
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/wallet:
    put:
      tags: [Auth]
      summary: Connect Wallet Address
      description: |
        Hubungkan wallet Polygon ke akun Aurora.
        Wallet address ini dipakai untuk fetch posisi real dari Polymarket Data API.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [walletAddress]
              properties:
                walletAddress:
                  type: string
                  pattern: '^0x[a-fA-F0-9]{40}$'
                  example: "0x742d35Cc6634C0532925a3b8D4C9B5D1b2c3D4E5"
      responses:
        '200':
          description: Wallet connected
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                      email:
                        type: string
                      walletAddress:
                        type: string
                  message:
                    type: string
                    example: "Wallet connected successfully"
        '400':
          description: Invalid wallet address format
        '401':
          description: Unauthorized
        '409':
          description: Wallet already connected to another account

  # ── Markets ─────────────────────────────────────────────────────────────────
  /api/markets:
    get:
      tags: [Markets]
      summary: Get Markets
      description: |
        Ambil daftar market dari DB (di-sync dari Polymarket Gamma API).
        Jika DB belum ada data, fallback fetch langsung dari Polymarket.
        Di-cache 60 detik.
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [Trending, New, Politics, Sports, Crypto]
          example: Crypto
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: search
          in: query
          schema:
            type: string
            minLength: 1
            maxLength: 100
          example: bitcoin
      responses:
        '200':
          description: Daftar market
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Market'
                  total:
                    type: integer
                    example: 20
        '400':
          description: Validation error
  /api/markets/{id}:
    get:
      tags: [Markets]
      summary: Get Market Detail
      description: Ambil detail satu market. ID bisa berupa conditionId, polymarketId, atau slug.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "will-btc-reach-100k-2026"
      responses:
        '200':
          description: Detail market
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Market'
        '404':
          description: Market not found
  /api/markets/{id}/price-history:
    get:
      tags: [Markets]
      summary: Get Price History
      description: |
        Data harga historis untuk chart di Aurora app.
        Format response disesuaikan dengan LineChart dari `react-native-gifted-charts`.
        Di-cache 5 menit (300 detik).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: interval
          in: query
          schema:
            type: string
            enum: [1m, 1h, 6h, 1d, 1w, all]
            default: 1d
          description: Resolusi data
        - name: outcome
          in: query
          schema:
            type: string
            default: Yes
          description: Nama outcome yang dicari (Yes, No, atau nama lain)
      responses:
        '200':
          description: Data price history
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PricePoint'
                  outcome:
                    type: string
                    example: "Yes"
                  tokenId:
                    type: string
                  interval:
                    type: string
                    example: "1d"
                  count:
                    type: integer
        '404':
          description: Market not found

  # ── Orderbook ───────────────────────────────────────────────────────────────
...